+++ packages/core/hardhat.config.js
@@ -29,7 +29,8 @@ module.exports = {
       { version: '0.5.16', settings },
       { version: '0.6.12', settings },
       { version: '0.7.6', settings },
-      { version: '0.8.2', settings },
+      { version: '0.8.8', settings },
+      { version: '0.8.9', settings },
     ],
   },
   etherscan: {